4# Create Area card with db

+++ app/src/main/java/com/example/ifplan_leite/data/entities/Animal.kt
@@ -5,7 +5,7 @@ import androidx.room.PrimaryKey
 
 @Entity(tableName = "animal_database")
 data class Animal(
-    @PrimaryKey(autoGenerate = true) val id: Long = 1,
+    @PrimaryKey(autoGenerate = true) val id: Int = 1,
     val pesoCorporal: Double,
     val milkProduction: Double,
     val milkFatContent: Double,
